% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_mp3.R
\name{cut_mp3}
\alias{cut_mp3}
\title{cut_mp3}
\usage{
cut_mp3(input_mp3, start, end, out_mp3 = NULL)
}
\arguments{
\item{input_mp3}{mp3 file path.}

\item{start}{cut start time point, format: HH:MM:SS(时分秒).}

\item{end}{cut end time point, format: HH:MM:SS(时分秒).}

\item{out_mp3}{output mp3 file path, default source file name plus time point.}
}
\value{
Nothing.
}
\description{
Cut mp3.
Cut mp3 with specific start and end time point.
}
\author{
Jiahao Wang
}
