% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formate_convert.R
\name{formate_convert}
\alias{formate_convert}
\title{formate_convert}
\usage{
formate_convert(from, to, files_input, out_name = NULL)
}
\arguments{
\item{from}{Source format, flv|mp4.}

\item{to}{Target format mp4|mp3.}

\item{files_input}{Input video path, one or more.}
}
\value{
Nothing.
}
\description{
Formate Convert for Video.
Only support convert flv to mp4 or mp4 to mp3.
}
\author{
Jiahao Wang
}
