% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_danmu.R
\name{get_danmu}
\alias{get_danmu}
\title{get_danmu}
\usage{
get_danmu(url, out_dir)
}
\arguments{
\item{url}{Video URL, required, remember with double quotation marks.}

\item{out_dir}{Output, directory, required.}
}
\value{
Nothing, infornation will be printed to screen.
}
\description{
Get danmaku of video.
Get danmaku of video.
}
\author{
Jiahao Wang
}
