% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{get_info}
\usage{
get_info(url)
}
\arguments{
\item{url}{Video URL, remember with double quotation marks.}
}
\value{
Nothing, infornation will be printed to screen.
}
\description{
Get information of video.
Get information of video.
}
\author{
Jiahao Wang
}
