% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video.R
\name{get_video}
\alias{get_video}
\title{get_video}
\usage{
get_video(
  url = NULL,
  out_dir = NULL,
  format = NULL,
  all_list = FALSE,
  out_name = NULL,
  mp3 = FALSE,
  start = NULL,
  end = NULL,
  remove.mp4 = FALSE,
  danmu = FALSE
)
}
\arguments{
\item{url}{Video URL, remember with double quotation marks.}

\item{out_dir}{Output directory, default current directory.}

\item{all_list}{Whether download videos in this vidio list(下载所有分P视频).}

\item{out_name}{Rename output file, DO NOT contain suffix, DO NOT use it when using 'all_list'.}

\item{mp3}{Whether convert mp4 to mp3.}

\item{start}{Start time point of mp3 cutting, format: HH:MM:SS(时分秒).}

\item{end}{End time point of mp3 cutting, format: HH:MM:SS(时分秒).}

\item{remove.mp4}{Whether remove mp4 file.}

\item{danmu}{Whether save danmaku.}
}
\value{
Nothing.
}
\description{
Download video using you-get.
Support Youtube/Bilibili and other non-VIP video.
}
\author{
Jiahao Wang
}
