% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_multi.R
\name{get_video_multi}
\alias{get_video_multi}
\title{get_video_multi}
\usage{
get_video_multi(
  config_input = NULL,
  out_dir = NULL,
  prepare = FALSE,
  prepare_name = "get_video_multi"
)
}
\arguments{
\item{config_input}{Input config file, not neccessary when build config using parameter 'prepare'.}

\item{out_dir}{Output directory, required.}

\item{prepare}{Automaticaly create and open a new config file.}

\item{prepare_name}{Name using for creat config file.}
}
\value{
Nothing.
}
\description{
Download video using you-get.
Download multiple videos with specific config file.
}
\author{
Jiahao Wang
After run get_video_multi(out_dir, prepare = TRUE), the config file with be created and open.
Values in column 'mp3 remove.mp4 danmu' should be assign with 1 if you want set to TRUE
And not assign angthing if you want set to FALSE and NULL
}
